requirements.txt
colour
Pillow
